Changed model wrapper name and moved models to appropriate folder. Added paper ref.
