+++ avalanche/models/__init__.py
@@ -13,3 +13,4 @@ from .mlp_tiny_imagenet import SimpleMLP_TinyImageNet
 from .mobilenetv1 import MobilenetV1
 from .dynamic_modules import *
 from .utils import *
+from slda_resnet import SLDAResNetModel
