+++ avalanche/training/strategies/deep_slda.py
@@ -3,67 +3,12 @@ from typing import Optional, Sequence
 
 import os
 import torch
-from torch import nn
-import torchvision.models as models
 
 from avalanche.training.plugins import StrategyPlugin
 from avalanche.training.strategies import BaseStrategy
 from avalanche.training import default_logger
 from avalanche.models.dynamic_modules import MultiTaskModule
-from avalanche.models import ModelWrapper
-
-
-class SLDAResNetModel(nn.Module):
-    """
-    This is a model wrapper to reproduce experiments from the original
-    paper of Deep Streaming Linear Discriminant Analysis by using
-    a pretrained ResNet model.
-    """
-
-    def __init__(self, arch='resnet18', output_layer_name='layer4.1',
-                 imagenet_pretrained=True, device='cpu'):
-        """
-        :param arch: backbone architecture (default is resnet-18, but others
-        can be used by modifying layer for
-        feature extraction in `self.feature_extraction_wrapper'
-        :param imagenet_pretrained: True if initializing backbone with imagenet
-        pre-trained weights else False
-        :param output_layer_name: name of the layer from feature extractor
-        :param device: cpu, gpu or other device
-        """
-
-        super(SLDAResNetModel, self).__init__()
-
-        feat_extractor = models.__dict__[arch](
-            pretrained=imagenet_pretrained).to(device).eval()
-        self.feature_extraction_wrapper = ModelWrapper(
-            feat_extractor, output_layer_name).eval()
-
-        warnings.warn(
-            "The Deep SLDA implementation is not perfectly aligned with "
-            "the paper implementation (i.e., it does not use a base "
-            "initialization phase here and instead starts streming from "
-            "pre-trained weights).")
-
-    @staticmethod
-    def pool_feat(features):
-        feat_size = features.shape[-1]
-        num_channels = features.shape[1]
-        features2 = features.permute(0, 2, 3,
-                                     1)  # 1 x feat_size x feat_size x
-        # num_channels
-        features3 = torch.reshape(features2, (
-            features.shape[0], feat_size * feat_size, num_channels))
-        feat = features3.mean(1)  # mb x num_channels
-        return feat
-
-    def forward(self, x):
-        """
-        :param x: raw x data
-        """
-        feat = self.feature_extraction_wrapper(x)
-        feat = SLDAResNetModel.pool_feat(feat)
-        return feat
+from avalanche.models import FeatureExtractorBackbone
 
 
 class StreamingLDA(BaseStrategy):
@@ -74,6 +19,8 @@ class StreamingLDA(BaseStrategy):
     The result is processed one element at a time to fit the
     LDA.
     Original paper:
+    "Hayes et. al., Lifelong Machine Learning with Deep Streaming Linear
+    Discriminant Analysis, CVPR Workshop, 2020"
     https://openaccess.thecvf.com/content_CVPRW_2020/papers/w15/Hayes_Lifelong_Machine_Learning_With_Deep_Streaming_Linear_Discriminant_Analysis_CVPRW_2020_paper.pdf
     """
     def __init__(self, slda_model, criterion,
@@ -90,7 +37,7 @@ class StreamingLDA(BaseStrategy):
         :param output_layer_name: if not None, wrap model to retrieve
             only the `output_layer_name` output. If None, the strategy
             assumes that the model already produces a valid output.
-            You can use `ModelWrapper` class to create your custom
+            You can use `FeatureExtractorBackbone` class to create your custom
             SLDA-compatible model.
         :param input_size: feature dimension
         :param num_classes: number of total classes in stream
@@ -110,7 +57,7 @@ class StreamingLDA(BaseStrategy):
             plugins = []
 
         if output_layer_name is not None:
-            slda_model = ModelWrapper(slda_model.to(device),
+            slda_model = FeatureExtractorBackbone(slda_model.to(device),
                                       output_layer_name).eval()
 
         super(StreamingLDA, self).__init__(
@@ -282,6 +229,5 @@ class StreamingLDA(BaseStrategy):
 
 
 __all__ = [
-    'StreamingLDA',
-    'SLDAResNetModel',
+    'StreamingLDA'
 ]
