+++ examples/deep_slda.py
@@ -25,9 +25,9 @@ from avalanche.training.plugins import EvaluationPlugin
 from avalanche.evaluation.metrics import accuracy_metrics, loss_metrics, \
     forgetting_metrics
 from avalanche.logging import InteractiveLogger
-from avalanche.benchmarks.classic import CORe50
-from avalanche.training.strategies.deep_slda import StreamingLDA, \
-    SLDAResNetModel
+from avalanche.benchmarks.classic import SplitCIFAR10
+from avalanche.training.strategies.deep_slda import StreamingLDA
+from avalanche.models import SLDAResNetModel
 
 
 def main(args):
@@ -50,8 +50,7 @@ def main(args):
     # ---------
 
     # --- SCENARIO CREATION
-    scenario = CORe50(scenario=args.scenario,
-                      train_transform=transform, eval_transform=transform)
+    scenario = SplitCIFAR10(5, train_transform=transform, eval_transform=transform)
     # ---------
 
     eval_plugin = EvaluationPlugin(
@@ -92,7 +91,7 @@ if __name__ == '__main__':
     parser.add_argument('--cuda', type=int, default=0,
                         help='Select zero-indexed cuda device. -1 to use CPU.')
 
-    parser.add_argument('--n_classes', type=int, default=50)
+    parser.add_argument('--n_classes', type=int, default=10)
     parser.add_argument('--scenario', type=str, default="nc",
                         choices=['ni', 'nc', 'nic', 'nicv2_79', 'nicv2_196',
                                  'nicv2_391'])
